refactor operation to top level
